Automatically backed up by Learn

+++ index.js
@@ -23,11 +23,11 @@ function dec(a) {
 }
 
 function makeInt(a) {
-  return parseInt('a', 10)
+  return parseFloat('a', 10)
 }
 
 
-/*
+
 describe('makeInt(n)', function() {
   it('parses n as an integer and returns the parsed integer', function() {
     expect(makeInt(a.toString())).toEqual(a)
